test: introduce systemTest target
While IntegrationTest allows to use the SpringContext,
SystemTest goes one level higher in the test pyramid
and runs using the docker containers.

Motivation: SystemTest take longer to run as the docker containers need to be started.
IntegrationTest strike a balance between fast to execute and high test coverage.
