+++ springwolf-examples/springwolf-amqp-example/src/test/java/io/github/stavshamir/springwolf/example/amqp/ProducerSystemTest.java
@@ -1,24 +1,31 @@
 // SPDX-License-Identifier: Apache-2.0
-package io.github.stavshamir.springwolf.example.sqs;
+package io.github.stavshamir.springwolf.example.amqp;
 
-import io.github.stavshamir.springwolf.example.sqs.consumers.ExampleConsumer;
-import io.github.stavshamir.springwolf.example.sqs.dtos.ExamplePayloadDto;
-import io.github.stavshamir.springwolf.producer.SpringwolfSqsProducer;
+import io.github.stavshamir.springwolf.example.amqp.consumers.ExampleConsumer;
+import io.github.stavshamir.springwolf.example.amqp.dtos.ExamplePayloadDto;
+import io.github.stavshamir.springwolf.producer.SpringwolfAmqpProducer;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
+import org.springframework.amqp.AmqpIOException;
+import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
+import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.TestPropertySource;
 import org.testcontainers.containers.DockerComposeContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 import java.io.File;
 
-import static io.github.stavshamir.springwolf.example.sqs.dtos.ExamplePayloadDto.ExampleEnum.FOO1;
+import static io.github.stavshamir.springwolf.example.amqp.dtos.ExamplePayloadDto.ExampleEnum.FOO1;
+import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
 import static org.mockito.Mockito.timeout;
 import static org.mockito.Mockito.verify;
 
@@ -27,23 +34,34 @@ import static org.mockito.Mockito.verify;
  * the setup uses a full docker-compose context with a real sqs instance.
  */
 @SpringBootTest(
-        classes = {SpringwolfSqsExampleApplication.class},
+        classes = {SpringwolfAmqpExampleApplication.class},
         webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Testcontainers
 @DirtiesContext
 @TestMethodOrder(OrderAnnotation.class)
+@TestPropertySource(properties = {"spring.rabbitmq.host=localhost"})
 // @Ignore("Uncomment this line if you have issues running this test on your local machine.")
-public class ProducerIntegrationWithDockerIntegrationTest {
+public class ProducerSystemTest {
 
     @Autowired
-    SpringwolfSqsProducer springwolfSqsProducer;
+    SpringwolfAmqpProducer springwolfAmqpProducer;
 
     @SpyBean
     ExampleConsumer exampleConsumer;
 
     @Container
     public static DockerComposeContainer<?> environment =
-            new DockerComposeContainer<>(new File("docker-compose.yml")).withServices("localstack");
+            new DockerComposeContainer<>(new File("docker-compose.yml")).withServices("amqp");
+
+    @Test
+    @Order(1)
+    void verifyAmqpIsAvailable() {
+        ConnectionFactory factory = new CachingConnectionFactory("localhost");
+
+        await().atMost(60, SECONDS).ignoreException(AmqpIOException.class).untilAsserted(() -> assertThat(
+                        factory.createConnection().isOpen())
+                .isTrue());
+    }
 
     @Test
     @Order(2)
@@ -55,7 +73,7 @@ public class ProducerIntegrationWithDockerIntegrationTest {
         payload.setSomeEnum(FOO1);
 
         // when
-        springwolfSqsProducer.send("example-queue", payload);
+        springwolfAmqpProducer.send("example-queue", payload);
 
         // then
         verify(exampleConsumer, timeout(10000)).receiveExamplePayload(payload);
