+++ springwolf-examples/springwolf-sqs-example/src/test/java/io/github/stavshamir/springwolf/example/sqs/ApiIntegrationTest.java
@@ -14,6 +14,8 @@ import org.springframework.test.context.DynamicPropertySource;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -38,7 +40,7 @@ class ApiIntegrationTest {
     void asyncApiResourceArtifactTest() throws JSONException, IOException {
         String url = "/springwolf/docs";
         String actual = restTemplate.getForObject(url, String.class);
-        System.out.println("Got: " + actual);
+        Files.writeString(Path.of("src", "test", "resources", "asyncapi.actual.json"), actual);
 
         InputStream s = this.getClass().getResourceAsStream("/asyncapi.json");
         String expected = new String(s.readAllBytes(), StandardCharsets.UTF_8);
